python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_backend"

[project]
name = "dployr-sdk"
version = "1.0.0"
description = "Python SDK for dployr"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Dployr", email = "hello@dployr.dev"}
]
keywords = ["dployr", "sdk", "api", "deployment"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "microsoft-kiota-abstractions>=1.3.0",
    "microsoft-kiota-http>=1.3.0",
    "microsoft-kiota-serialization-json>=1.0.0",
    "microsoft-kiota-serialization-text>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/dployr-io/dployr"
Documentation = "https://docs.dployr.dev"
Repository = "https://github.com/dployr-io/dployr-sdk"
Issues = "https://github.com/dployr-io/dployr/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["dployr_sdk*"]


